+++ src/collective/volto/sitesettings/restapi/deserializer/controlpanels.py
@@ -4,6 +4,7 @@ from collective.volto.sitesettings.interfaces import (
 from collective.volto.sitesettings.interfaces import (
     ICollectiveVoltoSitesettingsSiteControlpanel,
 )
+from plone import api
 from plone.formwidget.namedfile.converter import b64decode_file
 from plone.namedfile.file import NamedImage
 from plone.restapi.deserializer import json_body
@@ -26,6 +27,8 @@ except ImportError:
     # Plone 52
     from Products.CMFPlone.interfaces import ISiteSchema
 
+import json
+
 
 @implementer(IDeserializeFromJson)
 @adapter(ICollectiveVoltoSitesettingsSiteControlpanel)
@@ -37,6 +40,7 @@ class SiteSettingsDeserializeFromJson(ControlpanelDeserializeFromJson):
         not compromise other functionalities.
         """
         data = json_body(self.controlpanel.request)
+
         for schema_interface in [
             ISiteSchema,
             ICollectiveVoltoSitesettingsAdditionalSiteSchema,
@@ -86,7 +90,27 @@ class SiteSettingsDeserializeFromJson(ControlpanelDeserializeFromJson):
             if errors:
                 raise BadRequest(errors)
 
-        # set width and height for logo and favicon
+        self.align_site_title(data=data)
+        self.set_logo_and_favicon_sizes(data=data)
+
+    def align_site_title(self, data):
+        """align site_title with custom one"""
+        site_title_translated = data.get("site_title_translated", "")
+
+        if site_title_translated:
+            site_title_translated = json.loads(site_title_translated)
+        else:
+            site_title_translated = {}
+
+        lang = api.portal.get_default_language()
+        new_site_title = site_title_translated.get(lang, "")
+
+        if new_site_title:
+            proxy = self.registry.forInterface(ISiteSchema, prefix=self.schema_prefix)
+            setattr(proxy, "site_title", new_site_title.replace("\n", " "))
+
+    def set_logo_and_favicon_sizes(self, data):
+        """set width and height for logo and favicon"""
         proxy = self.registry.forInterface(
             ICollectiveVoltoSitesettingsAdditionalSiteSchema,
             prefix=self.schema_prefix,