+++ src/collective/volto/sitesettings/restapi/services/site/get.py
@@ -35,14 +35,14 @@ class Site(BaseSite):
             prefix="plone",
             check=False,
         )
+
         # set title and subtitle based on language, if field is set
-        title = {"default": result["site"]["plone.site_title"]}
-        title.update(self.json_to_dict(additional_settings, "site_title_translated"))
-
-        site_subtitle = self.json_to_dict(additional_settings, "site_subtitle")
-
-        result["site"]["plone.site_title"] = title
-        result["site"]["plone.site_subtitle"] = site_subtitle
+        result["site"]["plone.site_title"] = self.json_to_dict(
+            additional_settings, "site_title_translated"
+        )
+        result["site"]["plone.site_subtitle"] = self.json_to_dict(
+            additional_settings, "site_subtitle"
+        )
 
         # images
         site_url = api.portal.get().absolute_url()